color tracking
In this project, you'll use OpenCV (cv2) to detect and track a specific color in a video stream or from an image. Colour tracking can be used for various applications, such as object detection, motion tracking, or interactive systems . The goal of this project is to capture video input, identify a certain color range, and track the object of that color in real-time.